Update gungame.sma
Uploading as few changes made but not enough to constitute a version change. Hopefully optimized better for hook. When respawning after warm-up at times even without hook players gets stuck as they were using hook. Adjustment was made to make them the world instead of seperate ent as that created jitter amd sticking. Marked custom ent to be logged when it is dispatched as end_game to see if it is interfering with mapcycle to vote , "NEXT MAP", expectations. Also according to thread in forum if a certain line is not cut OP4 is crash-prone top bots. I left it in and it will disable itself if op4 is detected running. Why do bots need this anyway? Should be removed later I think. Needs testing.